feat : 메뉴 리팩터 및 실행 클래스 구축
