+++ src/main/java/org/application/InputView.java
@@ -1,0 +1,20 @@
+package org.application;
+
+import java.util.Scanner;
+
+public class InputView {
+
+    private static final Scanner scanner = new Scanner(System.in);
+
+    public static String inputMenuSelection() {
+
+        OutputView.outputByMenu();
+
+        return scanner.nextLine();
+    }
+
+    public static String inputOperation() {
+
+        return scanner.nextLine();
+    }
+}
