--- src/main/java/org/console/InputView.java
@@ -1,15 +1,0 @@
-package org.console;
-
-import java.util.Scanner;
-
-public class InputView {
-
-    private final Scanner scanner = new Scanner(System.in);
-
-    public String inputMenuSelection() {
-
-        OutputView.outputByMenu();
-
-        return scanner.nextLine();
-    }
-}
