+++ src/main/java/org/error/ResponseErrorFormat.java
@@ -4,6 +4,7 @@ public enum ResponseErrorFormat {
 
     ERROR_DIVISION_BY_ZERO("0으로는 나눌 수 없습니다."),
     ERROR_BAD_OPERATOR("잘못된 연산자입니다."),
+    ERROR_NOT_FOUND_MENU("없는 메뉴입니다."),
     ;
 
     private String message;
@@ -11,5 +12,9 @@ public enum ResponseErrorFormat {
     ResponseErrorFormat(String message) {
 
         this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
     }
 }
